Add raycast function to navmesh object
Implement in Python and AssemblyScript versions the new function raycast. It allows to find the intersection of the navigation mesh with a spatial ray. The ray is defined by origin and direction. The algorithm use bvh structure with triangles of the navigaton mesh. It tests the intersection of the aabb with the ray, and if it intersects then test the child (subnode or triangle).